Sockets in non blocking mode. Need to dive deeper into clients' sockets (created only with accept()?)
